requirements.txt

numpy
scipy
scikit-image
opencv-python
rasterio
geopandas
shapely
pyproj
rioxarray
xarray
matplotlib
tqdm
click
pyyaml
